removed setting migrate feature (fixes #24)
This feature is likely to cause more issues than it's worth. Settings incompatibilities between RespeQt and AspeQt are likely to grow in time, even for settings which are currently shared between the two. Besides that, the number of settings which need to be set in RespeQt are fairly small, and the additional work by the user to set them is rather small, and one-time.