Fix crash from double call to closeEvent
At least on OSX closeEvent gets called two times. This leads to invalid access crash on application exit.
This happes, because the pointer textPrinterWindow and docDisplayWindow are already released on second call, but the original values is still there.
The easy fix for this issue is setting the released pointers to NULL, so that delete doesn't crash.